+++ cli/setup/server.js
@@ -147,6 +147,7 @@ export async function createSetupServer({ packageRoot, projectRoot, port }) {
       }
     ],
     resolve: {
+      dedupe: ['vue'],
       alias: {
         '@setup': clientDir
       }